Added a GetToken function in axios.js
